requirements.txt
allure-pytest
allure-python-commons
attrs
certifi
cffi
charset-normalizer
colorama
et-xmlfile
exceptiongroup
h11
idna
iniconfig
openpyxl
outcome
packaging
Pillow
pluggy
pycparser
pytest
pytest-xdist
requests
selenium
sniffio
sortedcontainers
tomli
trio
trio-websocket
urllib3
wsproto
